requirements.txt
Django==4.1.13
environs==9.5.0
django-phonenumber-field==6.2.0
phonenumbers==8.12.50
pillow==10.3.0